int/labels: separates files for de/serialization and incrementing
